requirements.txt
fastapi==0.109.0
gensim==4.3.2
ipython==8.20.0
nltk==3.8.1
pandas==2.1.4
uvicorn==0.25.0
pydantic==2.5.3
PyYAML==6.0.1